Completed Api

+++ SampleProject/Models/Contact.cs
@@ -1,8 +1,4 @@
-using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations.Schema;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace SampleProject.Models
 {
@@ -10,7 +6,7 @@ namespace SampleProject.Models
    {
       public int Id { get; set; }
       public string Address { get; set; }
-      public int PhoneNumber { get; set; }
+      public long PhoneNumber { get; set; }
       public virtual Person Person { get; set; }
 
       [ForeignKey("Person")]
